fix: update import paths after parser and numStr files were moved
- buildInputData.js: ../parser/ → ./parser/ (parsers are now in same dir)
- parseOpenPositions.js: ../instrument/classifier.js → ../../../core/domain/instrument/classifier.js
- ThresholdWarning.jsx: ../../core/domain/numStr.js → @util/numStr.js
- HoldingsCalculator.js: ../parser/ → ../../../app/input/parser/
- calculateTax.js: ../domain/parser/ → ../../app/input/parser/
- inferPriorPositions.js: ../domain/parser/ → ../../app/input/parser/
- __tests__/calculateTax.js: ../taxService.js → ../calculateTax.js, mock paths updated
